Title: 입력받아 연산하기, Time: 6ms, Memory: 0MB, Status: Passed - Codetree

+++ 250407/입력받아 연산하기/take-input-and-operate.cpp
@@ -6,7 +6,7 @@ int main() {
     cin >> a >> b;
 
     a += 87;
-   double( b )%= 10;
+   b%= 10;
 
    cout << a << "\n" << b;
     return 0;
